Free all allocated memory for the queue
This function is responsible for freeing the memory used by the queue.
It first checks if the queue is NULL or empty to avoid unnecessary
operations. Then, during traversal, it removes each element’s prev and
next pointers from the list to ensure other nodes are not affected
before safely deleting the element and freeing its memory. Finally, it
frees the queue head to prevent memory leaks.

Change-Id: Id4a6f24cc59ca53bbcd03ec7bbb6c285ade0d1af